clean up room settings usage
